tickdata moved
