+++ BerlinProject/src/environments/reward_models.py
@@ -1,4 +1,4 @@
-from environments.tick_data import TickData
+from models.tick_data import TickData
 
 
 def reward_model_x(position: int, action: str, tick: TickData, buy_price:float) -> float:
