+++ BerlinProject/src/environments/simple_position.py
@@ -1,7 +1,7 @@
 import numpy as np
 from typing import List
 from environments.state import State
-from environments.tick_data import TickData
+from models.tick_data import TickData
 from config import BUY, SELL, ACTION
 
 HOLD_TIME_LIMIT = 20
