+++ BerlinProject/src/models/tick_data.py
