+++ BerlinProject/src/operations/chuckle_charts_web.py
@@ -4,11 +4,10 @@ import numpy as np
 from config.pyobject_id import PyObjectId
 from config.types import INDICATOR_COLLECTION
 from data_streamer.external_tool import ExternalTool
-from environments.tick_data import TickData
+from models.tick_data import TickData
 from models.monitor_configuration import MonitorConfiguration
 from models.monitor_model import Monitor
 from mongo_tools.mongo import Mongo
-from .web_socket_client import WebSocketClient
 from .restful_messager import RestfulMessenger
