+++ BerlinProject/tests/data_streamer/candle_aggregator_test.py
@@ -1,7 +1,6 @@
 import sys
 import os
 import logging
-from datetime import datetime
 from typing import List
 
 # Simple logging setup
@@ -18,7 +17,7 @@ if src_dir not in sys.path:
 from stock_analysis_ui.services.schwab_auth import SchwabAuthManager
 from data_streamer.schwab_data_link import SchwabDataLink
 from data_streamer.candle_aggregator import CandleAggregator
-from environments.tick_data import TickData
+from models.tick_data import TickData
 
 
 def test_prepopulation():
