+++ BerlinProject/tests/data_streamer/test_initialization_run.py
@@ -6,7 +6,7 @@ import logging
 from data_streamer.schwab_data_link import SchwabDataLink
 from data_streamer.data_streamer import DataStreamer
 from data_streamer.external_tool import ExternalTool
-from environments.tick_data import TickData
+from models.tick_data import TickData
 from models.monitor_configuration import MonitorConfiguration
