+++ BerlinProject/tests/test_mongo_tools/test_sample_tools.py
@@ -1,6 +1,6 @@
 import unittest
 
-from environments.tick_data import TickData
+from models.tick_data import TickData
 
 from mongo_tools.sample_tools import SampleTools
